Move BeaconBlockBody to a subpackage
