+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/datastructures/blocks/BeaconBlockBodyTest.java
@@ -24,6 +24,7 @@ import static tech.pegasys.teku.util.config.Constants.MAX_VOLUNTARY_EXITS;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.bls.BLSSignature;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
