+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/datastructures/blocks/BeaconBlockTest.java
@@ -22,6 +22,7 @@ import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
 class BeaconBlockTest {
